Move server HMR handling into hot-middleware
